syntax fixes

+++ lab2/train.py
@@ -23,11 +23,11 @@ def dense_to_one_hot(y, class_count):
 np.random.seed(int(time.time() * 1e6) % 2**31)
 
 ds_train, ds_test = MNIST(DATA_DIR, train=True, download=True), MNIST(DATA_DIR, train=False)
-train_x = ds_train.data.reshape([-1, 1, 28, 28]).numpy().astype(np.float) / 255
+train_x = ds_train.data.reshape([-1, 1, 28, 28]).numpy().astype(np.cfloat) / 255
 train_y = ds_train.targets.numpy()
 train_x, valid_x = train_x[:55000], train_x[55000:]
 train_y, valid_y = train_y[:55000], train_y[55000:]
-test_x = ds_test.data.reshape([-1, 1, 28, 28]).numpy().astype(np.float) / 255
+test_x = ds_test.data.reshape([-1, 1, 28, 28]).numpy().astype(np.cfloat) / 255
 test_y = ds_test.targets.numpy()
 train_mean = train_x.mean()
 train_x, valid_x, test_x = (x - train_mean for x in (train_x, valid_x, test_x))
